modificando comprobante grafico para su uso con el resto de ventanas
