--- src/gui/comprobante.java
@@ -1,462 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gui;
-
-
-/**
- * @author Andres Moises
- */
-public class comprobante extends javax.swing.JFrame {
-
-    /** Creates new form comprobante */
-    public comprobante(
-            DatosComprobante Datos) { // tira error pero es norma, tiene que recibir del archivo
-        initComponents();
-        datosFecha.setText(Datos.fecha);
-        datosCuentaDestino.setText(Datos.cuentaDestino);
-        datosMonto.setText(Datos.monto);
-        datosNumero.setText(Datos.numero);
-        datosConcepto.setText(Datos.concepto());
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form. WARNING: Do NOT
-     * modify this code. The content of this method is always regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        Comprobante = new javax.swing.JLabel();
-        Numero = new javax.swing.JLabel();
-        Fecha = new javax.swing.JLabel();
-        CuentaDestino = new javax.swing.JLabel();
-        Concepto = new javax.swing.JLabel();
-        Monto = new javax.swing.JLabel();
-        jButton1 = new javax.swing.JButton();
-        datosFecha = new javax.swing.JLabel();
-        datosCuentaDestino = new javax.swing.JLabel();
-        datosConcepto = new javax.swing.JLabel();
-        datosMonto = new javax.swing.JLabel();
-        datosNumero = new javax.swing.JLabel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        Comprobante.setFont(new java.awt.Font("Tahoma", 1, 18)); // NOI18N
-        Comprobante.setText("Comprobante");
-
-        Numero.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        Numero.setText("Fecha: ");
-
-        Fecha.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        Fecha.setText("Cuenta De Destino: ");
-
-        CuentaDestino.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        CuentaDestino.setText("Numero: ");
-
-        Concepto.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        Concepto.setText("Monto: ");
-
-        Monto.setFont(new java.awt.Font("Tahoma", 1, 14)); // NOI18N
-        Monto.setText("Concepto:");
-
-        jButton1.setText("Aceptar");
-        jButton1.addActionListener(
-                new java.awt.event.ActionListener() {
-                    public void actionPerformed(java.awt.event.ActionEvent evt) {
-                        jButton1ActionPerformed(evt);
-                    }
-                });
-
-        datosFecha.addAncestorListener(
-                new javax.swing.event.AncestorListener() {
-                    public void ancestorMoved(javax.swing.event.AncestorEvent evt) {}
-
-                    public void ancestorAdded(javax.swing.event.AncestorEvent evt) {
-                        datosFechaAncestorAdded(evt);
-                    }
-
-                    public void ancestorRemoved(javax.swing.event.AncestorEvent evt) {}
-                });
-
-        datosCuentaDestino.addAncestorListener(
-                new javax.swing.event.AncestorListener() {
-                    public void ancestorMoved(javax.swing.event.AncestorEvent evt) {}
-
-                    public void ancestorAdded(javax.swing.event.AncestorEvent evt) {
-                        datosCuentaDestinoAncestorAdded(evt);
-                    }
-
-                    public void ancestorRemoved(javax.swing.event.AncestorEvent evt) {}
-                });
-
-        datosConcepto.addAncestorListener(
-                new javax.swing.event.AncestorListener() {
-                    public void ancestorMoved(javax.swing.event.AncestorEvent evt) {}
-
-                    public void ancestorAdded(javax.swing.event.AncestorEvent evt) {
-                        datosConceptoAncestorAdded(evt);
-                    }
-
-                    public void ancestorRemoved(javax.swing.event.AncestorEvent evt) {}
-                });
-
-        datosMonto.addAncestorListener(
-                new javax.swing.event.AncestorListener() {
-                    public void ancestorMoved(javax.swing.event.AncestorEvent evt) {}
-
-                    public void ancestorAdded(javax.swing.event.AncestorEvent evt) {
-                        datosMontoAncestorAdded(evt);
-                    }
-
-                    public void ancestorRemoved(javax.swing.event.AncestorEvent evt) {}
-                });
-
-        datosNumero.addAncestorListener(
-                new javax.swing.event.AncestorListener() {
-                    public void ancestorMoved(javax.swing.event.AncestorEvent evt) {}
-
-                    public void ancestorAdded(javax.swing.event.AncestorEvent evt) {
-                        datosNumeroAncestorAdded(evt);
-                    }
-
-                    public void ancestorRemoved(javax.swing.event.AncestorEvent evt) {}
-                });
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                        .addGroup(
-                                layout.createSequentialGroup()
-                                        .addGroup(
-                                                layout.createParallelGroup(
-                                                                javax.swing.GroupLayout.Alignment
-                                                                        .LEADING)
-                                                        .addGroup(
-                                                                layout.createSequentialGroup()
-                                                                        .addGap(191, 191, 191)
-                                                                        .addGroup(
-                                                                                layout.createParallelGroup(
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .Alignment
-                                                                                                        .TRAILING)
-                                                                                        .addComponent(
-                                                                                                Numero)
-                                                                                        .addComponent(
-                                                                                                Fecha)
-                                                                                        .addComponent(
-                                                                                                Concepto)
-                                                                                        .addComponent(
-                                                                                                Monto)
-                                                                                        .addComponent(
-                                                                                                CuentaDestino))
-                                                                        .addGap(18, 18, 18)
-                                                                        .addGroup(
-                                                                                layout.createParallelGroup(
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .Alignment
-                                                                                                        .LEADING)
-                                                                                        .addComponent(
-                                                                                                datosFecha,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE,
-                                                                                                299,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE)
-                                                                                        .addComponent(
-                                                                                                datosCuentaDestino,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE,
-                                                                                                299,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE)
-                                                                                        .addComponent(
-                                                                                                datosConcepto,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE,
-                                                                                                299,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE)
-                                                                                        .addComponent(
-                                                                                                datosMonto,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE,
-                                                                                                299,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE)
-                                                                                        .addComponent(
-                                                                                                datosNumero,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE,
-                                                                                                299,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE)))
-                                                        .addGroup(
-                                                                layout.createSequentialGroup()
-                                                                        .addGap(347, 347, 347)
-                                                                        .addComponent(
-                                                                                Comprobante,
-                                                                                javax.swing
-                                                                                        .GroupLayout
-                                                                                        .PREFERRED_SIZE,
-                                                                                132,
-                                                                                javax.swing
-                                                                                        .GroupLayout
-                                                                                        .PREFERRED_SIZE)))
-                                        .addContainerGap(193, Short.MAX_VALUE))
-                        .addGroup(
-                                javax.swing.GroupLayout.Alignment.TRAILING,
-                                layout.createSequentialGroup()
-                                        .addGap(0, 0, Short.MAX_VALUE)
-                                        .addComponent(jButton1)
-                                        .addGap(87, 87, 87)));
-        layout.setVerticalGroup(
-                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                        .addGroup(
-                                layout.createSequentialGroup()
-                                        .addGap(49, 49, 49)
-                                        .addComponent(
-                                                Comprobante,
-                                                javax.swing.GroupLayout.PREFERRED_SIZE,
-                                                27,
-                                                javax.swing.GroupLayout.PREFERRED_SIZE)
-                                        .addGap(30, 30, 30)
-                                        .addGroup(
-                                                layout.createParallelGroup(
-                                                                javax.swing.GroupLayout.Alignment
-                                                                        .LEADING)
-                                                        .addComponent(
-                                                                CuentaDestino,
-                                                                javax.swing.GroupLayout
-                                                                        .PREFERRED_SIZE,
-                                                                30,
-                                                                javax.swing.GroupLayout
-                                                                        .PREFERRED_SIZE)
-                                                        .addComponent(
-                                                                datosNumero,
-                                                                javax.swing.GroupLayout
-                                                                        .PREFERRED_SIZE,
-                                                                30,
-                                                                javax.swing.GroupLayout
-                                                                        .PREFERRED_SIZE))
-                                        .addGap(30, 30, 30)
-                                        .addGroup(
-                                                layout.createParallelGroup(
-                                                                javax.swing.GroupLayout.Alignment
-                                                                        .TRAILING)
-                                                        .addGroup(
-                                                                layout.createSequentialGroup()
-                                                                        .addGroup(
-                                                                                layout.createParallelGroup(
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .Alignment
-                                                                                                        .TRAILING)
-                                                                                        .addGroup(
-                                                                                                layout.createSequentialGroup()
-                                                                                                        .addGroup(
-                                                                                                                layout.createParallelGroup(
-                                                                                                                                javax
-                                                                                                                                        .swing
-                                                                                                                                        .GroupLayout
-                                                                                                                                        .Alignment
-                                                                                                                                        .TRAILING)
-                                                                                                                        .addGroup(
-                                                                                                                                layout.createSequentialGroup()
-                                                                                                                                        .addGroup(
-                                                                                                                                                layout.createParallelGroup(
-                                                                                                                                                                javax
-                                                                                                                                                                        .swing
-                                                                                                                                                                        .GroupLayout
-                                                                                                                                                                        .Alignment
-                                                                                                                                                                        .TRAILING)
-                                                                                                                                                        .addComponent(
-                                                                                                                                                                Numero,
-                                                                                                                                                                javax
-                                                                                                                                                                        .swing
-                                                                                                                                                                        .GroupLayout
-                                                                                                                                                                        .PREFERRED_SIZE,
-                                                                                                                                                                30,
-                                                                                                                                                                javax
-                                                                                                                                                                        .swing
-                                                                                                                                                                        .GroupLayout
-                                                                                                                                                                        .PREFERRED_SIZE)
-                                                                                                                                                        .addComponent(
-                                                                                                                                                                datosFecha,
-                                                                                                                                                                javax
-                                                                                                                                                                        .swing
-                                                                                                                                                                        .GroupLayout
-                                                                                                                                                                        .PREFERRED_SIZE,
-                                                                                                                                                                30,
-                                                                                                                                                                javax
-                                                                                                                                                                        .swing
-                                                                                                                                                                        .GroupLayout
-                                                                                                                                                                        .PREFERRED_SIZE))
-                                                                                                                                        .addGap(
-                                                                                                                                                40,
-                                                                                                                                                40,
-                                                                                                                                                40)
-                                                                                                                                        .addComponent(
-                                                                                                                                                Fecha,
-                                                                                                                                                javax
-                                                                                                                                                        .swing
-                                                                                                                                                        .GroupLayout
-                                                                                                                                                        .PREFERRED_SIZE,
-                                                                                                                                                30,
-                                                                                                                                                javax
-                                                                                                                                                        .swing
-                                                                                                                                                        .GroupLayout
-                                                                                                                                                        .PREFERRED_SIZE))
-                                                                                                                        .addComponent(
-                                                                                                                                datosCuentaDestino,
-                                                                                                                                javax
-                                                                                                                                        .swing
-                                                                                                                                        .GroupLayout
-                                                                                                                                        .PREFERRED_SIZE,
-                                                                                                                                30,
-                                                                                                                                javax
-                                                                                                                                        .swing
-                                                                                                                                        .GroupLayout
-                                                                                                                                        .PREFERRED_SIZE))
-                                                                                                        .addGap(
-                                                                                                                37,
-                                                                                                                37,
-                                                                                                                37)
-                                                                                                        .addComponent(
-                                                                                                                Monto,
-                                                                                                                javax
-                                                                                                                        .swing
-                                                                                                                        .GroupLayout
-                                                                                                                        .PREFERRED_SIZE,
-                                                                                                                30,
-                                                                                                                javax
-                                                                                                                        .swing
-                                                                                                                        .GroupLayout
-                                                                                                                        .PREFERRED_SIZE))
-                                                                                        .addComponent(
-                                                                                                datosConcepto,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE,
-                                                                                                30,
-                                                                                                javax
-                                                                                                        .swing
-                                                                                                        .GroupLayout
-                                                                                                        .PREFERRED_SIZE))
-                                                                        .addGap(18, 18, 18)
-                                                                        .addComponent(
-                                                                                Concepto,
-                                                                                javax.swing
-                                                                                        .GroupLayout
-                                                                                        .PREFERRED_SIZE,
-                                                                                30,
-                                                                                javax.swing
-                                                                                        .GroupLayout
-                                                                                        .PREFERRED_SIZE))
-                                                        .addComponent(
-                                                                datosMonto,
-                                                                javax.swing.GroupLayout
-                                                                        .PREFERRED_SIZE,
-                                                                30,
-                                                                javax.swing.GroupLayout
-                                                                        .PREFERRED_SIZE))
-                                        .addGap(18, 18, 18)
-                                        .addComponent(jButton1)
-                                        .addContainerGap(53, Short.MAX_VALUE)));
-
-        pack();
-    } // </editor-fold>//GEN-END:initComponents
-
-    private void jButton1ActionPerformed(
-            java.awt.event.ActionEvent evt) { // GEN-FIRST:event_jButton1ActionPerformed
-        this.dispose();
-    } // GEN-LAST:event_jButton1ActionPerformed
-
-    private void datosFechaAncestorAdded(
-            javax.swing.event.AncestorEvent evt) { // GEN-FIRST:event_datosFechaAncestorAdded
-    } // GEN-LAST:event_datosFechaAncestorAdded
-
-    private void datosCuentaDestinoAncestorAdded(
-            javax.swing.event.AncestorEvent
-                    evt) { // GEN-FIRST:event_datosCuentaDestinoAncestorAdded
-        // TODO add your handling code here:
-    } // GEN-LAST:event_datosCuentaDestinoAncestorAdded
-
-    private void datosConceptoAncestorAdded(
-            javax.swing.event.AncestorEvent evt) { // GEN-FIRST:event_datosConceptoAncestorAdded
-        // TODO add your handling code here:
-    } // GEN-LAST:event_datosConceptoAncestorAdded
-
-    private void datosMontoAncestorAdded(
-            javax.swing.event.AncestorEvent evt) { // GEN-FIRST:event_datosMontoAncestorAdded
-        // TODO add your handling code here:
-    } // GEN-LAST:event_datosMontoAncestorAdded
-
-    private void datosNumeroAncestorAdded(
-            javax.swing.event.AncestorEvent evt) { // GEN-FIRST:event_datosNumeroAncestorAdded
-        // TODO add your handling code here:
-    } // GEN-LAST:event_datosNumeroAncestorAdded
-
-    static void mostrarComprobante() {
-        java.awt.EventQueue.invokeLater(
-                new Runnable() {
-                    public void run() {
-                        comprobante hola;
-                        hola = new comprobante();
-                        hola.setVisible(true);
-                    }
-                });
-    }
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        mostrarComprobante();
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JLabel Comprobante;
-    private javax.swing.JLabel Concepto;
-    private javax.swing.JLabel CuentaDestino;
-    private javax.swing.JLabel Fecha;
-    private javax.swing.JLabel Monto;
-    private javax.swing.JLabel Numero;
-    private javax.swing.JLabel datosConcepto;
-    private javax.swing.JLabel datosCuentaDestino;
-    private javax.swing.JLabel datosFecha;
-    private javax.swing.JLabel datosMonto;
-    private javax.swing.JLabel datosNumero;
-    private javax.swing.JButton jButton1;
-    // End of variables declaration//GEN-END:variables
-}
