+++ src/gui/menuPrincipal.java
@@ -1,3 +1,5 @@
 package gui;
 
-public class menuPrincipal {}
+public class menuPrincipal {
+
+}
